generator/image_gemini.py
"""Gemini image generation."""
import logging
from typing import Tuple, Optional
import google.generativeai as genai
import config

logger = logging.getLogger(__name__)

# Configure Gemini
genai.configure(api_key=config.GEMINI_API_KEY)


def generate_featured_image(
    topic: str,
    brand_name: str = "",
    language: str = "nl"
) -> Tuple[Optional[bytes], str, str]:
    """
    Generate a featured image using Gemini.
    
    Returns:
        Tuple of (image_bytes, mime_type, filename) or (None, "", "") on failure
    """
    try:
        # Build prompt - avoid trademarks/logos, produce clean blog header
        prompt = f"""Create a professional, clean blog header image for an article about: {topic}
        
Requirements:
- Modern, minimalist design
- No text, logos, or trademarks
- Suitable for blog featured image
- High quality, visually appealing
- Color scheme appropriate for {brand_name if brand_name else 'professional'} content
- Abstract or conceptual representation of the topic
"""
        
        # Use Gemini image generation
        # Note: Gemini 2.0 Flash may support image generation differently
        # This is a placeholder implementation - adjust based on actual Gemini API capabilities
        model = genai.GenerativeModel(config.GEMINI_IMAGE_MODEL)
        
        # For now, we'll use a text-to-image approach if available
        # If Gemini doesn't support direct image generation, we'll return None
        # and the job will continue without an image
        
        # Attempt to generate image (this may need adjustment based on actual Gemini API)
        try:
            response = model.generate_content(prompt)
            # If Gemini returns image data, extract it
            # Otherwise, return None to continue without image
            logger.warning("Gemini image generation not fully implemented - returning None")
            return None, "", ""
        except Exception as e:
            logger.warning(f"Gemini image generation failed: {e}. Continuing without image.")
            return None, "", ""
            
    except Exception as e:
        logger.error(f"Error generating image with Gemini: {e}")
        return None, "", ""


def generate_featured_image_fallback(
    topic: str,
    brand_name: str = "",
    language: str = "nl"
) -> Tuple[Optional[bytes], str, str]:
    """
    Fallback: Generate image using alternative method or return placeholder.
    For MVP, we'll skip image generation if Gemini doesn't support it directly.
    """
    # In a production system, you might:
    # 1. Use a different image generation service
    # 2. Use Unsplash API for stock images
    # 3. Use DALL-E or Midjourney API
    # For MVP, we'll just log and continue without image
    logger.info(f"Skipping image generation for topic: {topic}")
    return None, "", ""
