requirements.txt
Flask==3.0.0
pydantic==2.5.0
requests==2.31.0
cryptography==41.0.7
openai==1.58.1
google-generativeai==0.8.3
